+++ package.json
@@ -17,7 +17,6 @@
     "react": "^16.12.0",
     "react-bootstrap": "^1.0.0-beta.16",
     "react-dom": "^16.12.0",
-    "reactstrap": "^8.2.0",
     "socket.io-client": "^2.3.0"
   }
 }
