+++ pages/user/login.css
@@ -1,0 +1,30 @@
+html{
+    height: 100vh;
+}
+*{
+    margin: 0;
+    padding: 0;
+}
+.App{
+    border:1px solid black;
+    padding:10px;
+    border-radius: 12px;
+    box-shadow: 12px 10px 17px;
+    background-color: #eee;
+}
+.login{
+    display: flex;
+    align-items:  center;
+    height:96vh;
+    padding: 0;
+    margin: 0;
+    
+}
+.App button{
+    background: none;
+}
+.App input{
+    border:none;
+    border-bottom: 2px solid #04a6bf;
+    background: none;
+}
